requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-google-genai>=2.0.0
google-genai>=0.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
arxiv>=2.1.0
PyLaTeX>=1.4.2
openai>=1.0.0
sympy>=1.12
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
networkx>=3.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0
jinja2>=3.1.0 